# Serial and OMP optimization of EwaldRef