--- a/src/QMCHamiltonians/EwaldRef.h
+++ b/src/QMCHamiltonians/EwaldRef.h
@@ -448,8 +448,8 @@ real_t getKappaMadelung(real_t volume)
 {
   real_t radius = std::pow(3. * volume / (4 * M_PI), 1. / 3);
   //return 2 * M_PI / ( 8 * radius );
-  return 1.0;
-  //return radius / std::sqrt( 2 * M_PI );
+  //return 1.0;
+  return std::sqrt(M_PI) / radius;
 }
 
 /** Compute the Madelung constant to a given tolerance
@@ -495,8 +495,8 @@ real_t getKappaEwald(real_t volume)
 {
   real_t radius = std::pow(3. * volume / (4 * M_PI), 1. / 3);
   //return 2 * M_PI / ( 8 * radius );
-  return 1.0;
-  //return radius / std::sqrt( 2 * M_PI );
+  //return 1.0;
+  return radius / std::sqrt( 2 * M_PI );
 }
 
 /** Compute the Ewald interaction of a particle pair to a given tolerance
@@ -572,24 +572,30 @@ real_t ewaldSumTile(const RealMat& a,
   const real_t kfactor = 4 * M_PI / volume;
 
   real_t es(0);
-  for (size_t i = row_first; i < row_last; ++i)
-    for (size_t j = col_first; j < std::min(col_last, i); ++j)
-    {
-      RealVec r(R[i] - R[j]);
-      real_t qq = Q[i] * Q[j];
-      // Create real-/k-space fuctors for terms within the sums in formula 6
-      RspaceEwaldTerm rfunc(r, a, rconst);
+  {
+    ScopedTimer RspaceTimer(TimerManager.createTimer("Rspace"));
+    for (size_t i = row_first; i < row_last; ++i)
+      for (size_t j = col_first; j < std::min(col_last, i); ++j)
+      {
+        RealVec r(R[i] - R[j]);
+        real_t qq = Q[i] * Q[j];
+        // Create real-/k-space fuctors for terms within the sums in formula 6
+        RspaceEwaldTerm rfunc(r, a, rconst);
 
-      // Compute the constant term
-      real_t cval = -2 * M_PI * std::pow(kconv, 2) / volume;
-      // Compute the real-space sum (includes zero)
-      real_t rval = gridSum(rfunc, true, tol / qq);
+        // Compute the constant term
+        real_t cval = -2 * M_PI * std::pow(kconv, 2) / volume;
+        // Compute the real-space sum (includes zero)
+        real_t rval = gridSum(rfunc, true, tol / qq);
 
-      // Sum all contributions to get the Ewald pair interaction
-      es += qq * (cval + rval);
-    }
+        // Sum all contributions to get the Ewald pair interaction
+        es += qq * (cval + rval);
+      }
+  }
 
-  es += gridSumTile(R, Q, row_first, row_last, col_first, col_last, b, kconst, kfactor, tol);
+  {
+    ScopedTimer KspaceTimer(TimerManager.createTimer("Kspace"));
+    es += gridSumTile(R, Q, row_first, row_last, col_first, col_last, b, kconst, kfactor, tol);
+  }
   return es;
 }
 
@@ -610,8 +616,7 @@ real_t ewaldSumTile(const RealMat& a,
 real_t ewaldEnergy(const RealMat& a, const PosArray& R, const ChargeArray& Q, real_t tol = 1e-10)
 {
   // Timer for EwaldRef
-  NewTimer& EwaldTimer(*TimerManager.createTimer("EwaldRef"));
-  ScopedTimer totalEwaldTimer(&EwaldTimer);
+  ScopedTimer totalEwaldTimer(TimerManager.createTimer("EwaldRef"));
 
   // Number of particles
   const size_t N = R.size();
@@ -619,21 +624,24 @@ real_t ewaldEnergy(const RealMat& a, const PosArray& R, const ChargeArray& Q, re
   // Total Ewald potential energy
   real_t ve = 0.0;
 
-  // Maximum self-interaction charge product
-  real_t qqmax = 0.0;
-  for (size_t i = 0; i < N; ++i)
-    qqmax = std::max(std::abs(Q[i] * Q[i]), qqmax);
-
-  // Compute the Madelung term (Drummond 2008 formula 7)
-  real_t vm = madelungSum(a, tol * 2. / qqmax);
-
-  // Sum the Madelung self interaction for each particle
-  for (size_t i = 0; i < N; ++i)
-    ve += Q[i] * Q[i] * vm / 2;
-
-  if (true)
   {
-    ScopedTimer totalEwaldTimer(TimerManager.createTimer("EwaldSum"));
+    ScopedTimer totalMadelungTimer(TimerManager.createTimer("MadelungSum"));
+    // Maximum self-interaction charge product
+    real_t qqmax = 0.0;
+    for (size_t i = 0; i < N; ++i)
+      qqmax = std::max(std::abs(Q[i] * Q[i]), qqmax);
+
+    // Compute the Madelung term (Drummond 2008 formula 7)
+    real_t vm = madelungSum(a, tol * 2. / qqmax);
+
+    // Sum the Madelung self interaction for each particle
+    for (size_t i = 0; i < N; ++i)
+      ve += Q[i] * Q[i] * vm / 2;
+  }
+
+  if (false)
+  {
+    ScopedTimer EwaldSumTimer(TimerManager.createTimer("EwaldSum"));
     real_t ve_ewald(0);
 // Sum the interaction terms for all particle pairs
 #pragma omp parallel for reduction(+ : ve_ewald)
@@ -659,7 +667,7 @@ real_t ewaldEnergy(const RealMat& a, const PosArray& R, const ChargeArray& Q, re
   }
   else
   {
-    ScopedTimer totalEwaldTimer(TimerManager.createTimer("EwaldSumOpt"));
+    ScopedTimer EwaldSumTimer(TimerManager.createTimer("EwaldSumOpt"));
 
     size_t tile_size, max_n_tile_1d, n_tiles;
     if (N <= 1)