I took a stab at optimizing the performance of EwaldRef by focusing on serial optimization prior to adding threading.

As noted in #2169, the k-space part dominated the computational load in the original implementation.  By considering the asymptotic decay of the error function (erfc(x)->exp(-x^2)/x), you can write down a condition that balances the asymptotic decay of the real and k-space parts (same exponent).  This balancing condition leads to kappa=1.

Balancing the exponents results in the following speedup for the 32 atom NiO cell (run in serial on laptop):

| Exponents  | Time |
|--------------|------|
| Original  | 81.1845 s |
| Balanced   | 0.3007 s |

For the 128 atom cell, EwaldRef now takes about 8 seconds on a single core:

| Size |  Time |
|------|-------|
|  32  | 0.3007 s |
|  64  | 1.4828 s |
| 128  | 7.4832 s |

I next added OMP threading by first flattening the triangular charge and pair distance data into serial arrays (temporary N^2 memory cost) and then performing a single OMP reduction over the serialized, data parallel flat loop.

For the 128 atom problem, this resulted in the following times for each level of threading (note: there are 12 cores on the laptop):

| Threads |  Time |
|----|----|
|  1  |    7.6578 s |
|  2  |   3.8986 s |
|  4  |    2.1098 s |
|  8   |   1.5173 s |
| 16   |   1.2456 s |

A few final adjustments are needed following discussion prior to merge.