@prckent 2019-12-30
Terrific speedup.

@jtkrogel, @ye-luo  Given #2172 what do you suggest? Can the optimized kappa choice be added to Ye's recently merged code? Or is the code in this PR preferred?