@jtkrogel 2019-12-30
It could, though as a reference implementation the comparative simplicity of the code in this PR is an advantage and probably a 2 second cost is good enough.  

Also, the prior threading optimizations targeted the k-space part which should now be roughly balanced with the real-space portion.