@ye-luo 2019-12-31
I followed @jtkrogel 's direction and found the balance between R and K parts by making the exp() and erfc() balance at the same large index shell (ix, iy, iz).
For the MadelungSum, the optimal choice is sqrt(pi)/Radius
For the EwaldSum, the optimal choice is Radius/sqrt(2 pi).
The cost of EwaldSum is N_atom^2 of MadelungSum. The original code picked a good choice for MadelungSum which is terrible for EwaldSum at large cells.
I cannot obtain the kappa = 1.0 as @jtkrogel but my choice gives me another factor of 2 speed up. After connecting the tiling code, 16 threads Xeon gives me 0.0457 second with the 128 atom cell.
```
      EwaldSumOpt                      0.0456     0.0237              1       0.045625925
        Kspace                         0.0010     0.0010              1       0.000994921
        Rspace                         0.0209     0.0209              1       0.020933867
```
The time difference between K and R part is due to the fact that I only optimized the K part.
Because the R and K are balanced now, I'd like to also optimize the R part (separated PR) to simplify the current code paths, although the performance improvements should be unimportant.