@ye-luo 2019-12-31
I also tried the 512 atom cell, 20 sec with kappa = 1.0 and after my change, 0.7 sec.
The tiling optimization code path is separated from the base code path.
@jtkrogel  `if (false)` code path in the function ewaldEnergy is the base code path and can be used to explore other schemes without interference from my optimization.