@jtkrogel 2019-12-31
Sorry to be slow on the reply.  I was looking more into optimal kappa yesterday and I came to basically the same conclusion as @ye-luo for kappa.

Here are updated results of the simply parallelized code (single omp parallel for):

| size | kappa | threads=1 | threads=16 |
|----|----|----|----|
| 128  | 1.0  |  7.6578 s  |  1.2456 s |
| 128  | opt  |  0.3805 s |  0.0638 s |
| 256  | opt  |  1.4934 s  |  0.2363 s |
| 512  | opt  |  7.1191 s  |  1.1219 s |

I get 0.06 s for 128 and 1.12 s for 512 w/ 16 threads on my mobile Xeon with the simply parallelized code.  This compares with  0.05 s  for 128 and 0.7 s for 512 that Ye finds with the more complex code on his (non-mobile?) Xeon.

The simple code is as fast as the more complex code.  Also, any further optimizations below 1 second are not meaningful here.  

I strongly prefer using simpler code as the reference.  If all agree, I will make a new PR with the simplified code + optimal kappa.