@anbenali 2019-12-31
Impressive!

On Tue, Dec 31, 2019, 15:46 Jaron T. Krogel <notifications@github.com>
wrote:

> Closed #2176 <https://github.com/QMCPACK/qmcpack/pull/2176>.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/QMCPACK/qmcpack/pull/2176?email_source=notifications&email_token=AF4XVQ4J4T6L5COD2573QCTQ3O4R5A5CNFSM4KBOYLE2YY3PNVWWK3TUL52HS4DFWZEXG43VMVCXMZLOORHG65DJMZUWGYLUNFXW5KTDN5WW2ZLOORPWSZGOVXIKCCY#event-2916131083>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AF4XVQY2DNLNBUWTMQVNUC3Q3O4R5ANCNFSM4KBOYLEQ>
> .
>