@ye-luo 2019-12-31
The current PR is simple. We only need to flip the false to true. I will handle the cleaning of the optimized code path. No need to create a new PR.