@ye-luo 2020-01-01
By the way, in the initial commits, flattening the triangle loop needs extra Staging memory for qq and rr. The simple code path under `if (false)` uses a simple parallel for without extra memory.