@ye-luo 2020-01-01
@prckent The code is as light as before adding the optimized code.