@jtkrogel 2020-01-02
I added some comments and restored the simpler threading.  This is just as fast and more transparent.

There is a temporary memory cost to flattening the triangular loop, but even for a 1024 atom system it amounts to 2M real values or about 16 MB.  If this code is ever adapted for production use, the use of these temporary arrays will be eliminated by having access to distance tables.

Ready for merge.  I will next consider the anisotropy intrinsic to the quasi-2D vdW systems.